feat : 특정시점 가격 조회

+++ src/main/java/com/aswemake/api/aswemakeapitask/dto/item/response/ItemPriceAtTimeResponseDto.java
@@ -14,6 +14,6 @@ public class ItemPriceAtTimeResponseDto {
     private Long id;
     private String name;
     private String date;
-    private int price;
-    private int currentPrice;
+    private Long price;
+    private Long currentPrice;
 }
